PathSize

Description

Since: 0.1

Class

class Monad m => MonadPathSize m where

Typeclass for finding a path's recursive size.

Since: 0.1

Methods

findLargestPaths

Arguments

:: HasCallStack

=> Config
Configuration.

-> Path
Path to search.

-> m (PathSizeResult SubPathData)
The results.

Given a path, finds the size of all subpaths, recursively.

Since: 0.1

Types

data PathData a

Associates a path to its total (recursive) size in the file-system. Takes a parameter for the size, allowing us to use a more efficient temporary type (e.g. Integer) before finally converting to Natural.

Since: 0.1

Constructors

MkPathData

Fields

path :: !Path

Path.

Since: 0.1

size :: !a

Size in bytes.

Since: 0.1

numFiles :: !a

Number of files.

Since: 0.1

numDirectories :: !a

Number of directories.

Since: 0.1

data SubPathData where

A flattened and sorted representation of PathTree. Contains at least one element.

Since: 0.1

Bundled Patterns

pattern MkSubPathData :: NESeq (PathData Natural) -> SubPathData
Pattern synonym for SubPathData. Note that construction sorts the underlying NESeq, so it is not constant.

Since: 0.1

data PathSizeResult a

Result of running a path-size computation with multiple notions of failure.

Since: 0.1

Constructors

PathSizeSuccess !a
Successfully computed a.

Since: 0.1

PathSizePartial !(NESeq PathE) !a
Computed a with some errors.

Since: 0.1

PathSizeFailure !(NESeq PathE)
Failed computing a.

Since: 0.1

Configuration

data Config

Since: 0.1

Constructors

MkConfig

Fields

searchAll :: !Bool

Whether to search hidden files/directories.

Since: 0.1

maxDepth :: !(Maybe Natural)

The depth limit of our search. Note that we still need to fully traverse the file system to get accurate data; this argument merely affects what is reported i.e. any depths > d are implicitly included in parent directories, but not directly.

exclude :: !(HashSet FilePath)

Paths to skip.

Since: 0.1

filesOnly :: !Bool

Whether to limit our search to just files.

Since: 0.1

numPaths :: !(Maybe Natural)

The number of paths to return.

Since: 0.1

strategy :: !Strategy

The search strategy.

Since: 0.1

data Strategy

Describes the path search strategy.

Since: 0.1

Constructors

Sync
Since: 0.1

Async
Since: 0.1

AsyncPooled
Since: 0.1

High level functions

pathSizeRecursive :: MonadPathSize m => Path -> m (PathSizeResult Natural)

Returns the total path size in bytes. Calls pathSizeRecursiveConfig with the following config:

MkConfig
  { searchAll = True,
    maxDepth = Just 0,
    exclude = mempty,
    filesOnly = False,
    numPaths = Just 1,
    strategy = mempty
  }

Since: 0.1

pathSizeRecursiveConfig :: MonadPathSize m => Config -> Path -> m (PathSizeResult Natural)

Returns the total path size in bytes.

Since: 0.1

display :: Bool -> SubPathData -> Text

Displays the data.

Since: 0.1

Errors

data PathE

Exception for a path. The second param is the reason i.e. the exceptions' displayException. The reason we convert to a string rather than leave it as an exception is so we can have an NFData instance for benchmarking.

Since: 0.1

Constructors

MkPathE !Path !String