PathSize.Exception

Description

Provides exception types/functions for usage with path-size.

Since: 0.1

Documentation

data PathE

Exception for a path. The second param is the reason i.e. the exceptions' displayException. The reason we convert to a string rather than leave it as an exception is so we can have an NFData instance for benchmarking.

Since: 0.1

Constructors

MkPathE !OsPath !String