PathSize.Utils

Description

Provides utilities.

Since: 0.1

Documentation

tryCalcSymLink :: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPosixCompat m) => OsPath -> m (PathSizeResult PathTree)

tryCalcFile :: (HasCallStack, MonadCatch m, MonadPathReader m) => OsPath -> m (PathSizeResult PathTree)

unzipResultSeq :: Seq (PathSizeResult PathTree) -> (Seq PathE, Seq PathTree)

Unzips a sequence of results.

Since: 0.1

hidden :: OsPath -> Bool

Detects hidden paths via a rather crude dot check, with an exception for the current directory ./.

Since: 0.1