PathSize.Data.Config

Description

Supplies PathSize Config.

Since: 0.1

Documentation

data Config

Since: 0.1

Constructors

MkConfig

Fields

searchAll :: !Bool

Whether to search hidden files/directories.

Since: 0.1

maxDepth :: !(Maybe Natural)

The depth limit of our search. Note that we still need to fully traverse the file system to get accurate data; this argument merely affects what is reported i.e. any depths > d are implicitly included in parent directories, but not directly.

exclude :: !(HashSet FilePath)

Paths to skip.

Since: 0.1

filesOnly :: !Bool

Whether to limit our search to just files.

Since: 0.1

numPaths :: !(Maybe Natural)

The number of paths to return.

Since: 0.1

strategy :: !Strategy

The search strategy.

Since: 0.1

data Strategy

Describes the path search strategy.

Since: 0.1

Constructors

Sync
Since: 0.1

Async
Since: 0.1

AsyncPooled
Since: 0.1

_Sync :: Prism' Strategy ()

_Async :: Prism' Strategy ()

_AsyncPooled :: Prism' Strategy ()

Since: 0.1