PathSize.Data.SubPathData

Description

Provides SubPathData type.

Since: 0.1

Type

data SubPathData where

A flattened and sorted representation of PathTree. Contains at least one element.

Since: 0.1

Bundled Patterns

pattern MkSubPathData :: NESeq PathData -> SubPathData
Pattern synonym for SubPathData. Note that construction sorts the underlying NESeq, so it is not constant.

Since: 0.1

Construction

mkSubPathData :: Bool -> PathTree -> SubPathData

Creates a SubPathData from a PathTree.

Since: 0.1

takeLargestN :: HasCallStack => Bool -> Positive Int -> PathTree -> SubPathData

Retrieves the largest N paths.

Since: 0.1

Elimination

unSubPathData :: SubPathData -> NESeq PathData

Since: 0.1

Operations

display :: Bool -> SubPathData -> Text

Displays the data.

Since: 0.1

Reexports

data NESeq a infixr 5

A general-purpose non-empty (by construction) finite sequence type.

Non-emptiness means that:

Functions that take an NESeq can safely operate on it with the assumption that it has at least value.

Functions that return an NESeq provide an assurance that the result has at least one value.

Data.Sequence.NonEmpty re-exports the API of Data.Sequence, faithfully reproducing asymptotics, typeclass constraints, and semantics. Functions that ensure that input and output maps are both non-empty (like <|) return NESeq, but functions that might potentially return an empty map (like tail) return a Seq instead.

You can directly construct an NESeq with the API from Data.Sequence.NonEmpty; it's more or less the same as constructing a normal Seq, except you don't have access to empty. There are also a few ways to construct an NESeq from a Seq:

The nonEmptySeq smart constructor will convert a Seq a into a Maybe (NESeq a), returning Nothing if the original Seq was empty.

You can use :<||, :||>, and insertSeqAt to insert a value into a Seq to create a guaranteed NESeq.

You can use the IsNonEmpty and IsEmpty patterns to "pattern match" on a Seq to reveal it as either containing a NESeq or an empty sequence.

withNonEmpty offers a continuation-based interface for deconstructing a Seq and treating it as if it were an NESeq.

You can convert an NESeq into a Seq with toSeq or IsNonEmpty, essentially "obscuring" the non-empty property from the type.