PathSize.Utils

Description

Provides utilities.

Since: 0.1

Windows / Unix compat

type MonadPosixC (m :: Type -> Type) = MonadPosix m

Alias for MonadPosix* constraints. On Posix, this is MonadPosix (unix), which allows for greater efficiency. On Windows, this is just MonadPosixCompat (unix-compat).

hidden :: OsPath -> Bool

Detects hidden paths via a rather crude dot check, with an exception for the current directory ./.

Since: 0.1

getFileStatus :: (HasCallStack, MonadPosixC m) => OsPath -> m FileStatus

Retrieves the FileStatus for the given path.

Since: 0.1

Misc

unzipResultSeq :: Seq (PathSizeResult PathTree) -> (Seq PathE, Seq PathTree)

Unzips a sequence of results.

Since: 0.1